requirements.txt
autopep8 == 1.5.3
Django == 3.0.7
djangorestframework == 3.11.0
django-cors-headers == 3.4.0
psycopg2 == 2.8.5
django-extensions == 2.2.9
python-dateutil == 2.8.1
djangorestframework-camel-case == 1.2.0
djangorestframework-jwt == 1.11.0
